+++ src/sink_input_effects.cpp
@@ -104,10 +104,10 @@ SinkInputEffects::SinkInputEffects(PulseManager* pulse_manager) : PipelineBase("
   pm->sink_input_removed.connect(sigc::mem_fun(*this, &SinkInputEffects::on_app_removed));
   pm->sink_changed.connect(sigc::mem_fun(*this, &SinkInputEffects::on_sink_changed));
 
-  g_settings_bind(settings, "buffer-out", source, "buffer-time", G_SETTINGS_BIND_DEFAULT);
-  g_settings_bind(settings, "latency-out", source, "latency-time", G_SETTINGS_BIND_DEFAULT);
-  g_settings_bind(settings, "buffer-out", sink, "buffer-time", G_SETTINGS_BIND_DEFAULT);
-  g_settings_bind(settings, "latency-out", sink, "latency-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "buffer-pulsesrc", source, "buffer-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "latency-pulsesrc", source, "latency-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "buffer-pulsesink", sink, "buffer-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "latency-pulsesink", sink, "latency-time", G_SETTINGS_BIND_DEFAULT);
 
   g_settings_bind(settings, "blocksize-out", adapter, "blocksize", G_SETTINGS_BIND_DEFAULT);