+++ src/source_output_effects.cpp
@@ -71,10 +71,10 @@ SourceOutputEffects::SourceOutputEffects(PulseManager* pulse_manager) : Pipeline
   pm->source_output_removed.connect(sigc::mem_fun(*this, &SourceOutputEffects::on_app_removed));
   pm->source_changed.connect(sigc::mem_fun(*this, &SourceOutputEffects::on_source_changed));
 
-  g_settings_bind(settings, "buffer-in", source, "buffer-time", G_SETTINGS_BIND_DEFAULT);
-  g_settings_bind(settings, "latency-in", source, "latency-time", G_SETTINGS_BIND_DEFAULT);
-  g_settings_bind(settings, "buffer-in", sink, "buffer-time", G_SETTINGS_BIND_DEFAULT);
-  g_settings_bind(settings, "latency-in", sink, "latency-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "buffer-pulsesrc", source, "buffer-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "latency-pulsesrc", source, "latency-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "buffer-pulsesink", sink, "buffer-time", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind(child_settings, "latency-pulsesink", sink, "latency-time", G_SETTINGS_BIND_DEFAULT);
 
   g_settings_bind(settings, "blocksize-in", adapter, "blocksize", G_SETTINGS_BIND_DEFAULT);